Add try/catch throw blocks to TwoListSynchronizer - only aiding debugging.
I am attempting to reflect a potentially multimap to multimap scenario; if I select one member of a multimap I want to modify the ItemsSelected to include all members of the multimap on the same side and also select all items on the other side of the map. This is difficult as I essentially want to modify the selected items during the CollectionChanged event, which is not a good idea. Additions to the Demo model are a thought WIP, incomplete.